Refactor file processing functions into `files` module
Moved certificate creation, file conversion, PDF processing, and combination logic from `audit.py` to the `files` module. This improves modularity and separates concerns for easier maintenance and scalability.
